+++ src/app/api/products/[productId]/variants/[variantId]/route.ts
@@ -9,12 +9,59 @@ import {
 } from "@/src/lib/types";
 import { NextRequest } from "next/server";
 import { NextResponse } from "next/server";
+import { validateToken, hasPermissions } from "@/src/dal/utils";
 
 export async function PUT(
   request: NextRequest,
   { params }: { params: Promise<{ productId: string; variantId: string }> },
 ): Promise<NextResponse<UpdateVariantResponse>> {
   try {
+    const authHeader = request.headers.get("Authorization");
+
+    if (!authHeader || !authHeader.startsWith("Bearer ")) {
+      return new NextResponse(
+        JSON.stringify({
+          message: "No token provided",
+        }),
+        {
+          status: 401,
+          headers: { "Content-Type": "application/json" },
+        },
+      );
+    }
+
+    const token = authHeader.split(" ")[1];
+    const authPayload = await validateToken(token);
+
+    if (!authPayload) {
+      return new NextResponse(
+        JSON.stringify({
+          message: "Invalid token",
+        }),
+        {
+          status: 401,
+          headers: { "Content-Type": "application/json" },
+        },
+      );
+    }
+
+    const isAuthorized = hasPermissions(
+      authPayload.permissions as Array<string>,
+      "update:variants",
+    );
+
+    if (!isAuthorized) {
+      return new NextResponse(
+        JSON.stringify({
+          message: "Unauthorized",
+        }),
+        {
+          status: 401,
+          headers: { "Content-Type": "application/json" },
+        },
+      );
+    }
+
     const body: ProductVariantUpdateDTO = await request.json();
 
     const newVariant = await updateProductVariant({
@@ -53,6 +100,52 @@ export async function DELETE(
   { params }: { params: Promise<{ productId: string; variantId: string }> },
 ) {
   try {
+    const authHeader = request.headers.get("Authorization");
+
+    if (!authHeader || !authHeader.startsWith("Bearer ")) {
+      return new NextResponse(
+        JSON.stringify({
+          message: "No token provided",
+        }),
+        {
+          status: 401,
+          headers: { "Content-Type": "application/json" },
+        },
+      );
+    }
+
+    const token = authHeader.split(" ")[1];
+    const authPayload = await validateToken(token);
+
+    if (!authPayload) {
+      return new NextResponse(
+        JSON.stringify({
+          message: "Invalid token",
+        }),
+        {
+          status: 401,
+          headers: { "Content-Type": "application/json" },
+        },
+      );
+    }
+
+    const isAuthorized = hasPermissions(
+      authPayload.permissions as Array<string>,
+      "delete:variants",
+    );
+
+    if (!isAuthorized) {
+      return new NextResponse(
+        JSON.stringify({
+          message: "Unauthorized",
+        }),
+        {
+          status: 401,
+          headers: { "Content-Type": "application/json" },
+        },
+      );
+    }
+
     const deletedVariant = await deleteProductVariant(
       (await params).productId,
       (await params).variantId,
